refactor(fantasy): move foundation files + extract mappers to feature module
- Copy types.ts, constants.ts, helpers.ts verbatim to src/features/fantasy/
- Replace originals with re-exports for backward compatibility
- Extract deriveEventStatus + dbEventToFantasyEvent into mappers/eventMapper.ts
- Extract dbHoldingToUserDpcHolding into mappers/holdingMapper.ts
- Clean up unused imports (DbEvent, HoldingWithPlayer, EventStatus, LineupFormat) from FantasyContent.tsx
