+++ src/app/(app)/fantasy/FantasyContent.tsx
@@ -13,7 +13,6 @@ import { useWallet } from '@/components/providers/WalletProvider';
 import { centsToBsd } from '@/lib/services/players';
 import { lockEventEntry, unlockEventEntry } from '@/lib/services/events';
 import { useUserTickets } from '@/lib/queries/tickets';
-import type { HoldingWithPlayer } from '@/lib/services/wallet';
 import { submitLineup, getLineup } from '@/lib/services/lineups';
 import { getFixtureDeadlinesByGameweek, getGameweekStatuses } from '@/lib/services/fixtures';
 import type { FixtureDeadline } from '@/lib/services/fixtures';
@@ -24,10 +23,9 @@ import { mapErrorToKey, normalizeError } from '@/lib/errorMessages';
 import { useEvents, useJoinedEventIds, usePlayerEventUsage, useHoldingLocks, useLeagueActiveGameweek, useIsClubAdmin } from '@/lib/queries/events';
 import { useHoldings } from '@/lib/queries/holdings';
 import { fmtScout, cn } from '@/lib/utils';
-import type { DbEvent } from '@/types';
 import {
-  type EventStatus, type FantasyTab, type FantasyEvent,
-  type LineupPlayer, type UserDpcHolding, type LineupFormat,
+  type FantasyTab, type FantasyEvent,
+  type LineupPlayer, type UserDpcHolding,
   CreateEventModal, SpieltagTab,
 } from '@/components/fantasy';
 import { getFormationsForFormat, buildSlotDbKeys } from '@/components/fantasy/constants';
@@ -36,6 +34,8 @@ import { MitmachenTab } from '@/components/fantasy/MitmachenTab';
 import { ErgebnisseTab } from '@/components/fantasy/ErgebnisseTab';
 import { EventsTab } from '@/components/fantasy/EventsTab';
 import { ScoringRules } from '@/components/fantasy/ScoringRules';
+import { dbEventToFantasyEvent } from '@/features/fantasy/mappers/eventMapper';
+import { dbHoldingToUserDpcHolding } from '@/features/fantasy/mappers/holdingMapper';
 
 import { useClub } from '@/components/providers/ClubProvider';
 import EventSummaryModal, { isEventSeen, markEventSeen } from '@/components/fantasy/EventSummaryModal';
@@ -62,102 +62,6 @@ const EventDetailModal = dynamic(
     ),
   }
 );
-
-// ============================================
-// MAPPERS: DB → Local Types
-// ============================================
-
-/** Derive actual event status from DB status — admin-controlled, no timestamp overrides */
-function deriveEventStatus(db: DbEvent): EventStatus {
-  if (db.scored_at) return 'ended';
-  const s = db.status;
-  if (s === 'ended' || s === 'scoring') return 'ended';
-  if (s === 'running') return 'running';
-  if (s === 'registering' || s === 'late-reg') return s as EventStatus;
-  return 'upcoming';
-}
-
-/** Map DB event to local FantasyEvent shape */
-function dbEventToFantasyEvent(db: DbEvent, joinedIds: Set<string>, userLineup?: { total_score: number | null; rank: number | null; reward_amount: number } | null): FantasyEvent {
-  return {
-    id: db.id,
-    name: db.name,
-    description: `${db.name} – ${db.format} Format`,
-    type: db.type === 'special' ? 'special' : db.type,
-    mode: db.format === '11er' ? 'tournament' : 'league',
-    status: deriveEventStatus(db),
-    format: (db.format || '7er') as LineupFormat,
-    gameweek: db.gameweek ?? 1,
-    startTime: new Date(db.starts_at).getTime(),
-    endTime: db.ends_at ? new Date(db.ends_at).getTime() : new Date(db.starts_at).getTime() + 259200000,
-    lockTime: new Date(db.locks_at).getTime(),
-    buyIn: db.currency === 'scout' ? centsToBsd(db.ticket_cost ?? db.entry_fee) : (db.ticket_cost ?? 0),
-    entryFeeCents: db.ticket_cost ?? db.entry_fee,
-    prizePool: centsToBsd(db.prize_pool),
-    guaranteed: centsToBsd(db.prize_pool),
-    participants: db.current_entries,
-    maxParticipants: db.max_entries,
-    entryType: 'single',
-    speed: 'normal',
-    sponsorName: db.sponsor_name ?? undefined,
-    sponsorLogo: db.sponsor_logo ?? undefined,
-    isPromoted: db.type === 'bescout' || db.type === 'sponsor',
-    isFeatured: db.type === 'sponsor',
-    isJoined: joinedIds.has(db.id),
-    isInterested: false,
-    userRank: userLineup?.rank ?? undefined,
-    userPoints: userLineup?.total_score ?? undefined,
-    userReward: userLineup?.reward_amount ?? undefined,
-    scoredAt: db.scored_at,
-    eventTier: db.event_tier ?? 'club',
-    minSubscriptionTier: db.min_subscription_tier ?? null,
-    salaryCap: db.salary_cap ? centsToBsd(db.salary_cap) : null,
-    minScPerSlot: db.min_sc_per_slot ?? 1,
-    wildcardsAllowed: db.wildcards_allowed ?? false,
-    maxWildcardsPerLineup: db.max_wildcards_per_lineup ?? 0,
-    requirements: { dpcPerSlot: db.min_sc_per_slot ?? 1 },
-    rewards: [
-      { rank: '1st', reward: 'Champion Badge' },
-      { rank: 'Top 10', reward: 'Gold Frame' },
-    ],
-    rewardStructure: db.reward_structure ?? null,
-    scope: db.scope ?? 'global',
-    lineupSize: db.lineup_size ?? (db.format === '11er' || db.format === '11er-reserve' ? 11 : 7),
-    ticketCost: db.ticket_cost ?? 0,
-    currency: db.currency ?? 'tickets',
-    clubId: db.club_id ?? undefined,
-    clubName: (db as Record<string, unknown>).clubs ? ((db as Record<string, unknown>).clubs as { name: string; logo_url: string | null }).name : undefined,
-    clubLogo: (db as Record<string, unknown>).clubs ? ((db as Record<string, unknown>).clubs as { name: string; logo_url: string | null }).logo_url ?? undefined : undefined,
-  };
-}
-
-/** Map DB holding row to local UserDpcHolding shape */
-function dbHoldingToUserDpcHolding(h: HoldingWithPlayer): UserDpcHolding {
-  return {
-    id: h.player_id,
-    first: h.player?.first_name ?? '',
-    last: h.player?.last_name ?? '',
-    pos: h.player?.position ?? 'MID',
-    club: h.player?.club ?? '',
-    clubId: h.player?.club_id ?? null,
-    dpcOwned: h.quantity,
-    eventsUsing: 0,
-    dpcAvailable: h.quantity,
-    activeEventIds: [],
-    isLocked: false,
-    lastScore: h.player?.perf_l5 ?? undefined,
-    avgScore: h.player?.perf_l15 ?? undefined,
-    perfL5: h.player?.perf_l5 ?? 0,
-    perfL15: h.player?.perf_l15 ?? 0,
-    matches: h.player?.matches ?? 0,
-    goals: h.player?.goals ?? 0,
-    assists: h.player?.assists ?? 0,
-    status: (h.player?.status as UserDpcHolding['status']) ?? 'fit',
-    imageUrl: h.player?.image_url ?? null,
-    ticket: h.player?.shirt_number ?? 0,
-    floorPrice: h.player?.floor_price ?? 0,
-  };
-}
 
 // ============================================
 // MAIN ORCHESTRATOR
